chore: moves firebase into other folder
