+++ src/config/zendesk/features.ts
@@ -1,5 +1,5 @@
 import { ZendeskAPI } from "react-zendesk";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 
 export const ZendeskOpenChat = () => {
   ZendeskAPI("webWidget", "open");
