+++ src/hooks/modalHooks/useAnimationReceiveTicketModal/index.tsx
@@ -5,7 +5,7 @@ import useVoucher from "hooks/useVoucher";
 import Ticket from "assets/icons/ticket-rounded-icon.svg";
 import SupportersIcon from "assets/icons/community-icon.svg";
 import UserIcon from "assets/icons/user-mono-icon.svg";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 import { useCausesContext } from "contexts/causesContext";
 import { useCurrentUser } from "contexts/currentUserContext";
 import { isFirstAccess } from "lib/onboardingFirstAccess";
