+++ src/layouts/LayoutHeader/ChangeLanguageItem/index.tsx
@@ -3,7 +3,7 @@ import CardIconText from "components/moleculars/cards/CardIconText";
 import ButtonSwitch from "components/atomics/buttons/ButtonSwitch";
 import { useLanguage } from "hooks/useLanguage";
 import { useTranslation } from "react-i18next";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 import * as S from "./styles";
 
 function ChangeLanguageItem(): JSX.Element {
