+++ src/layouts/LayoutHeader/index.tsx
@@ -17,7 +17,7 @@ import { useIntegrationId } from "hooks/useIntegrationId";
 import useNavigation from "hooks/useNavigation";
 import { useBlockedDonationModal } from "hooks/modalHooks/useBlockedDonationModal";
 import { RIBON_COMPANY_ID } from "utils/constants";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 import { useModal } from "hooks/modalHooks/useModal";
 import ChangeLanguageItem from "./ChangeLanguageItem";
 import LogoutItem from "./LogoutItem";
