+++ src/pages/donations/StoriesPage/index.tsx
@@ -6,7 +6,7 @@ import Story from "types/entities/Story";
 import useNavigation from "hooks/useNavigation";
 import { useIntegrationId } from "hooks/useIntegrationId";
 import useIntegration from "hooks/apiHooks/useIntegration";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 import useUsers from "hooks/apiHooks/useUsers";
 import useSources from "hooks/apiHooks/useSources";
 import { useCurrentUser } from "contexts/currentUserContext";
