+++ src/pages/promoters/SupportCausePage/CardPage/index.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useEffect, useState } from "react";
-import { logEvent } from "services/analytics";
+import { logEvent } from "services/analytics/firebase";
 import useCauses from "hooks/apiHooks/useCauses";
 import Cause from "types/entities/Cause";
 import IntersectBackground from "assets/images/intersect-background.svg";
