+++ src/services/analytics/firebase/index.ts
@@ -1,9 +1,24 @@
 import firebase from "firebase/app";
 import "firebase/analytics";
-import { logError } from "../crashReport";
+import { logError } from "../../crashReport";
 
 interface EventParams {
   [key: string]: string | number | undefined;
+}
+
+export function initializeFirebase(): any {
+  if (process.env.NODE_ENV === "development") return;
+  const firebaseConfig = {
+    apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+    authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+    projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
+    storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
+    messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
+    appId: process.env.REACT_APP_FIREBASE_APP_ID,
+    measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
+  };
+
+  firebase.initializeApp(firebaseConfig);
 }
 
 export function convertParamsToString(params: EventParams): EventParams {
